Moved display connector to near board edge
Also moved the display position silkscreen a bit, but the final location
is TBD pending construction of the Rev A prototype. As such, the current
3D model of the display doesn't really line up so well anymore.

The idea is that because the display needs to be raised a bit above the
PCB for improved visibility, this position change will allow for that
flexibility. Since its hard to accurately measure and remodel the
display location at this stage, and because the enclosure top is the
easiest component to tweak/recreate, its position will be measured from
the first build and fed back into the silkscreen later.