Sensor gain calibration
Implemented the sensor clean startup process and read loop.
Implemented various helper functions for running the sensor.
Added code to handle saving/loading settings using the built-in EEPROM.
Implemented a process for figuring out the sensor gain settings
to make basic count calculations more accurate.

The gain calibration process is necessary to ensure that we can
accurately compare sensor readings taken at different gain levels. The
datasheet defines each gain setting as a range, so we'd like to pin that
down to more exact values for our specific device. So far results seem
fairly close to the datasheet's "typical" figures, so it remains to be
seen how necessary this is.