Initial implementation of reflection measurements
Implementation of measurement and calibration is still being done
through the CDC protocol handler. Eventually it will be refactored to
somewhere else, and also hooked up to a local UI.

Also started to add a separate document to describe the CDC serial
protocol, so that there is a reference outside of the code itself. It
will be expanded and fleshed out over time.