Initial implementation of transmission density
Calibration for transmission is based on the "basic count" sensor values
for no film and a CAL-HI target, as well as the known density of that
CAL-HI target.

It may make sense to switch to storing reflection density as a "basic
count" value (rather than a log10 of such) too, simply to allow for more
flexibility in calculation. However, that will be left to later.

There's still room for improvement in the sensor read loop, as it
sometimes has glitches with the startup process on high density targets.