Add initial support for detect switch
Since the updated enclosure now allows the detect switch (BTN5) to work
correctly, we need some initial software support for it. As this switch
does not currently act like as an event trigger, like the other buttons,
it will be handled differently for now.

- Disabled interrupt support for BTN5
- Added functions to wrap polling the BTN5 state
- Added a BTN5 indicator on the diagnostic screen
- Updated the display states so that BTN5 toggles the idle-state read
  lights on and off, and gates whether or not measurements can be taken

If the detect switch isn't contacting correctly, then the device will
behave as though its always in the "pushing down" position. This was our
only state before, so it won't cause any real problems.