Implement basic measurement-zero support
It is now possible to set the current reflection or transmission
measurement as the "zero" point, to make it easier to measure things
relative to a film/paper base.

To set a zero point, hold down the measure button and press "down" after
the reading is complete.
To clear a zero point, hold down the measure button and press "up" after
the reading is complete.

It should be possible to set and/or clear in the display state, but
doing so will require tracking long-press duration. The current
simplistic key event handling is getting a bit too tangled to do that
with ease.