Initial refresh of CubeMX for the STM32L072 switch
This copies over the settings from our previous CubeMX config, moving
pins around when necessary. So far only the button GPIO pins have had to
change location.

STM's USB device stack is enabled to ensure proper USB code generation,
however the actual firmware uses TinyUSB instead.

FreeRTOS is also enabled, as we plan to start using this in the firmware
as one of the benefits of switching to an MCU with more flash headroom.

Refs #8