Changed MCU on main board and routed pin changes
Swapped the MCU from the L052K to the K072K, pulling in all pin
assignment changes from the CubeMX data.

Added extra bypass capacitor and sensor interrupt (/w pull-up) to
schematic.

Updated PCB routing to reflect changes, which included rearranging the
layout of many bypass capacitors and making a variety of adjustments to
routing in the MCU area.

Left a placeholder for a BOOT0 selector button that will be added in the
future.

Note: Decided to stick with KiCad's stock QFN-32 footprint instead of
using the custom one (based on the ST datasheet) used in the previous
revision. This is because the pads on the KiCad footprint extend further
out from the part, which should resolve some issues with edge solder
contact that happened during prototype assembly.

Refs #8