Add boot mode selection button
This adds a small right-angle push button, intended to be pressed via a
hole in the enclosure, to trigger the USB DFU bootloader. The current
design is pulled down with a 10k resistor normally, but pulled up to +5V
for the mode switch. There's also a 330 ohm in-line resistor like the
other buttons.

This design is subject to change, but its hard to find good references
for exactly what we're trying to do here.

Refs #11