Moved boot selector button again
Clarified schematic to show location of TVS diode as being right next to
the button, where it actually is.

Moved the boot selector button to the other side of the board, where it
is going to be more "out of the way" from the user perspective. Also,
moving it away from the Tag-Connect footprint makes routing feel a
little less crunched. The only downside is that the new location is over
the +5V power plane, so both signal and power traces need to be routed
away from the button to elsewhere on the board.

Refs #11