Migrate project from Rev A hardware to Rev B hardware support
This is a basic first pass, which doesn't involve any functional
changes of note.

Imported the CubeMX code generation dump, which includes changes for:
- Switching from the L052 to the L072 MCU
- Necessary GPIO pin assignment changes as part of the switch
- Adding FreeRTOS to the build, and making any necessary adjustments
  (e.g. moving HAL off systick)

Disabled -flto again for now, as we've gained a lot of headroom.

While FreeRTOS is now included in the build, it is not used yet in any
functional way. Migration of runtime behavior over to FreeRTOS tasks
will happen in future commits.