Initial FreeRTOS task-based startup integration
Created a central module to handle starting FreeRTOS tasks for all major
functions. This puts all the task definitions in one place, and starts
tasks in sequence using a semaphore to keep initialization behavior
predictable.

The currently-implemented tasks include:
- Main Task
  - Has the bulk of the module startup code that used to be in main()
  - Starts all the other tasks
  - Runs the state controller loop, and is generally responsible for
    interactive device behavior from that point onward
- USBD task
  - Background process of the tinyUSB library
- CDC task
  - Handles the command interface that uses a USB virtual serial port

Other tasks will be added as implemented, likely for keypad and sensor
functions.

In addition, quite a bit of hook/diagnostic code was enabled to make it
easier to debug FreeRTOS issues. This includes basic fault handler
hooks, and an improved hard fault handler. This will need to be cleaned
up at some point in the future.

In the process of doing all this work, it was discovered that printf()
has malloc-related  side-effects are very bad on a system with limited heap
memory. Since we're trying to give all of that heap over to the FreeRTOS
domain, there isn't much left to be haphazardly swallowed for these
purposes.  As such, direct printf() calls are now avoided at all costs.
If they are ever needed in the future, an alternative implementation
will need to be incorporated.

Thankfully, all the other functions we use from the printf family do not
appear to have any issues.

More cleanup

fixup