Keypad task integration
This change moves us away from the old model of polling keypad state
(using interrupts to detect changes) over to a fancier approach of using
a task and queues to handle key events.

It is based on the key handling routines from the Printalyzer, but
simplified because we are still reading key state directly from GPIOs.

This new approach makes it possible to treat key actions as actual
events, and block a task while waiting for a key press. It also makes it
possible to implement key repeat, which will be very useful for some
cases of using the up/down buttons.

Some notes on the integration:
- Mashing keys too quickly could cause the raw queues to overflow, which
led to missed events and sometimes failed timer assertions. The solution
was to specifically detect cases of trying to add events to a full raw
key event queue, and provide special handling of them.

- Button enum values now renamed to be based on function.
Given that the device graphic overlay now designates all these buttons
to specific standard roles, the code would be more clear if they were
internally named that way too.

- Only up/down buttons have any need to repeat.
Allowing repeat on any other buttons will just lead to spastic UI
behavior.

- Made keypad task logging optional.
Enabling logging on a task can add up to 512B of additional needed stack
size. Better to have it off by default on utility tasks, or we'll run
out of memory really quick.

- Detect button is routed through the same key interrupt handling and
event queue process as everything else. However, its value is still
read via directly polling the GPIO. The reason for this is that we care
about its current state more than the last state change, and don't want
to miss its current state in overflow situations.