Implemented sensor processing as a task
There is still plenty more work to do in terms of cleaning up the rough
edges, but the sensor process is now running as a task.

The task works by monitoring a queue to which commands and interrupt
events are sent.  Right now the only commands are to start, stop, and
set the gain/time config. These commands also use a semaphore to allow
the main task to block on their execution.

Interrupt events trigger a sensor read, and the updating of a value in a
1-element queue. This makes it possible to block on the next sensor
reading, without having to worry about the queue filling up with old
values.

The densitometer and sensor calibration functions have been rewritten to
work off this task, which dramatically reduces the amount of boilerplate
TSL2591 calls that were previously scattered around.

Still need to figure out a good way of communicating errors from the
sensor task back up to the main task, as that is still a TBD.  There's
also an open question of where dynamic behaviors like auto-gain are best
implemented.

The various sensor-oriented commands in the CDC handler will now need to
be rewritten, as there is currently nothing stopping them from
conflicting with main task behavior.

Also fixed a bug where the wrong max-gain calibration value was being
stored.