Major sensor task refactoring and LED cal integration
Split the sensor module into separate files, where one contains the
sensor task code and the other contains data types and utility functions
for interacting with it.

Moved responsibility for LED control into the sensor task, giving the
sensor task the ability to very tightly control the timing of LED
activation in relation to the sensor's integration cycle.

Added a calibration routine that runs the LED through a full 2-minute on
cycle, measures its brightness over time, then does logarithmic
regression to find a curve that models the change in brightness. This
will be used in a future commit to make the general sensor read loop
capable of compensating for brightness drop across multiple
measurements. In testing this seems to be a much bigger factor for
transmission than for reflection, but it is still implemented for both.

Removed the integration time calibration routines and settings, because
we are always taking operational measurements at the same (~200ms)
integration time setting. Calibrating this adds extra hassle, and only
has benefit if we wanted to compare readings across different time
settings.

Updated the gain calibration time to be more aware of LED
warmup/cooldown issues, which gets more accurate results.

Added cancellation ability to all sensor calibration callbacks, as well
as sensitivity to the detect switch.

There is still probably more cleanup work to do in the CDC handler,
however that task will be handled in the future.