Added more tusb callback handling and write mutex
Added the tusb cdc callback for RX, which releases a semaphore that the
main cdc task loop now blocks on. This removes the need to put a small
idle delay in that loop, so it can behave more correctly.

Added a write mutex to make it possible for code outside of the cdc
handler to send data out without conflicting.

Still need to work on handling the case of a disconnected terminal.