Add experimental "abort transfer" function to CDC stack code
This is an attempt at clearing transmit data that has already been
flushed, but hasn't actually been received by the host yet.
It may not be the best way of doing this, but it seems to work, so its
better than nothing for now.

Will remove/change if the tinyUSB project ever provides a better way of
dealing with the situation.