Transmission LED board revision B update
This change is to test adding a substantial thermal pad around the LED,
to see if it has any measurable effect on the brightness drop.

If the test is successful, a similar footprint change will likely be
made to the main PCB for its next revision. If the test is unsuccessful,
this revision may simply be rolled back and forgotten about.