Disable and/or remove LED calibration code paths
The first attempt at LED profiling may work well over a 30+ second
cycle, but is quite inaccurate and potentially causes problems within
the shorter period in which we do actual measurements. For now its safer
to completely rip it out until it can be reevaluated.

The underlying calibration routines are being left in, reachable only
via the CDC interface, as a way of comparing different LED
configurations of the hardware. Even those may be removed eventually.

Ultimately, the chosen solution is going to need to fall into one of two
categories:
1. A much fancer rethinking of how to profile LED behavior, possibly
   involving polynomial interpolation across multiple test runs.
2. A change to the target reading process, that simply ensures all
   readings are taken during the exact same time period after the LEDs
   are turned on, regardless of auto gain behavior.

Option 2 is probably going to be a lot easier to get right, and if
combined with better thermal handling of the LEDs (as attempted on
transmission PCB rev B), it could be good enough.