Initial app descriptor and checksum implementation
Added an app descriptor block to the application and its build process,
along with a checksum that is validated on firmware startup.

Deciding whether to actually do the validation, or how to react to a
validation failure, is still somewhat TBD. By the point in the startup
process where the validation code is running, a fair bit of the system
initialization has already been attempted. However, not enough
initialization has been performed to actually display any error message
to a user.

It might make sense to simply remove any error halting in favor of just
setting a flag that's made visible elsewhere if anyone cares. However,
for now, this is good enough.

Fixes #15