The great USB-CDC command protocol rewrite
Added data structures and utility functions for handling commands and
command responses in a new standardized format.

Reimplemented the majority of the old commands within the structure of
the new format, omitting a few that no longer made sense to retain.

Implemented a special "remote control" state for the UI state controller
- Several CDC commands will conflict with the local UI. To make them work,
  we need a way to put the device into a special state that locks out
  normal user interaction.
- This remote control state doesn't currently do anything useful on its
  own. There is an expectation that the CDC handler will do any real work
  during this state. If that's not viable, perhaps due to stack limitations,
  then inter-task communication can be added as needed.
- This remote control state is entered and exited via a new task
  notification mechanism, that lets the CDC handler explicitly tell the
  state controller what to do next.

Shuffled stack sizes between USBD and CDC tasks
- The USBD task wasn't using all of its allocated stack space, and there
  were a few cases where the CDC task needed just a tad more than it had.

Left in a few TODO comments that indicate next steps for certain
commands.

Still need to update the documentation of the new protocol