Update sensor read loop to standardize on-to-read time
To manage LED brightness drop-off effects, the amount of time from
turning the LED on to taking a reading should be constant, regardless of
what needs to happen in the measurement loop.

One way of doing this is by assuming we'll only ever need two gain
settings (high and max), and that we can pick our gain independently
from taking our actual readings.

Refs #14