Added ability to set slope calibration and changed sensor cal UI
Added CDC commands and settings to help with setting slope calibration
values.

The sensor calibration (gain and slope) are mostly intended for
"factory" use, but they're still somewhat user visible.

Changed the gain calibration UI to show the current values, with the
option to calibrate them hidden at the bottom of the list.

Added a slope calibration UI that just shows the currently set values,
as these are always calibrated as part of the desktop app.
Doing a full slope calibration requires filling out a form with 21+
known density values and their corresponding basic count sensor
readings. This is way too cumbersome to do on-device, so we're not even
going to try.

Refs #12