Change gain detection for the sensor target read function
Doing gain detection based on a 100ms reading, then switching to 200ms
for measurements had one corner case.  The saturation point at 100ms was
slightly greater than half the 200ms range, so it was possible to have
cases where we'd stick with max gain and then fail on saturation during
the measurement read.

Since the code in this function is tightly coupled to the integration
time values now, and because these values are unlikely to change
casually, their constants were also removed.