Remove support for non-HEX-encoded floating point commands
With the exception of the basic density reading output format, the
USB-CDC interface is intended for use with a dedicated desktop
application or utility. There is no longer any need to keep this
interface usable by a human who does not have access to such a utility.

The encoded hex format is much easier to handle consistently than
human-readable floating point formats, and requires less library code
too. As such, for the sake of firmware simplicity, this will be the
only format supported for most of the commands in the USB-CDC protocol.