Implemented a UF2 bootloader to ease the firmware update process
This bootloader is substantially derived from the TinyUF2 project, and
tries very hard to fit within 16k of space at the start of flash.
To achieve this, it has to be compiled with -flto and it cannot have
display output enabled at the same time as UART-based debug logging.

The display output implementation is extremely minimal, and contains all
of its image assets in the form of data already prepared to be written
directly to the display. No graphics library is used here to make things
easier, so the best way of preparing this data is to modify a u8g2
implementation to log its raw SPI writes and then capture them.

Right now the display code only shows two indicators. One indicates
being in bootloader mode, and the other indicates firmware load
progress. It may make sense to add additional indicicators in the
future, as space permits.

The code that handles writing to flash it written around assumptions of
the behavior of the UF2 implementation. Specifically, it assumes that
data will always be written in blocks that are a multiple of the flash
page size. This allows the implementation to use a much simpler and less
error-prone erase/write implementation than the TinyUF2 project examples
have.

There is still room for performance improvements in flash writing,
perhaps attempting the half-page approach laid out in the STM32L0x2
reference manual. However, that will be left to a future commit.

The function that jumps to the app code needs ito be tagged with the
__USED attribute, to prevent -flto from messing with it. Without this
attribute, sometimes this function causes a mysterious hard fault.
Adding it fixes the problem and makes it work reliably.
Some cleanup and flash_erase work (WIP)

The main device firmware was also modified slightly to support the
addition of the bootloader:
- Changed its linker script to the new flash allocation
- Enabled necessary vector table relocation
- Removed startup CRC check, because the bootloader is now doing CRC
  verification prior to starting the main firmware