Initial commit of companion desktop software
This desktop software is very much a work-in-progress, however it now
has the basic functions necessary to interact with the device.
It can connect, receive diagnostic information, capture screenshots, and
aid in calibration.

Its most important function is helping with the slope calibration process,
which cannot be done on-device.

There is still a ton of work and a fair bit of cleanup necessary before
this can be considered a generally usable application, however it is
time to get the basics in place. This is especially important since the
current iteration of the firmware's USB-CDC protocol handling was
developed in tandem with this application.

Refs #17