Updated footprints for LQFP-32 and USB-C connector
LQFP-32 footprint is based on the stock footprint, modified according to
the recommended footprint in ST's datasheet. This reduces the length of
the pads under the leads, which may help reduce bridging.

USB-C footprint is based on the stock Palcon footprint, modified
according to the GCT datasheet. This is several very minor changes, that
might not even make much of a difference.