Update schematic properties
Changed all revision labeling to "C" with today's date. This is subject
to change when the boards are sent for manufacturing, but removing
placeholders make mistakes here less likely in the future.

Set grid and drill origins to what they're hopefully supposed to be. Not
sure how much this matters, but it probably won't hurt.

More layout nitpicking and review changes may happen between now and
pre-production, but the general design and layout for Rev C can
probably be considered sufficiently done for now.