Added initial use of the IWDG watchdog to the system
At this stage, we are mainly adding this to prevent any system freezes
caused by spurious events or faults. It is unlikely that more complex
software problems, such as deadlocks, will be caught by the current
implementation.

The bootloader now initializes the watchdog with a 2 second timeout,
and kicks it in a few critical places. Once the main firmware is
started, it reduces that to a more sensible 0.5s period and kicks it
inside the FreeRTOS idle hook.

As such, this will only catch cases where the system goes directly into
an infinite loop due to a fault, or a task prevents the FreeRTOS
scheduler from running. In rare cases where a long-running tight loop is
expected (such as display screenshot capture), the watchdog is
explicitly kicked.

Also added the logging of reset reason flags to the startup proces, in
case there is anything we want to do with these in the future.

Since the CSR register is cleared as part of the shutdown code that runs
in the bootloader immediately prior to application jump, its value needs
to be saved elsewhere. The easiest place to put it is one of the RTC backup
registers, since we may also end up storing other information here as
well. Eventually this should also be used to distinguish whether the
reset was caused by bootloader or firmware behavior.

Refs #23