Add bulk capacitor to VBUS input
Added a 4.7uF capacitor just past the VBUS input ferrite bead, before
the +5V line goes to the 5V power plane.
Decided to stick with 4.7uF instead of 10uF for component commonality
and to avoid going past 10uF when including the LDO input capacitor.

Changed the 4.7uF capacitor used in both areas of the design to one with
a 25V rating, just to be safe.

Updated PCB layout and nudged some silkscreen bits around.

Fix #26