Update CubeMX specs to reassign LED enable pins
The original assignments for the LED enable pins (TLED_EN and RLED_EN)
are used by USART2 when the system bootloader is invoked.

This means that all of our measurement LEDs would switch on at full
power, if the device was booted into USB DFU mode for out-of-band
firmware updates.

While not technically a problem for normal users, allowing this does
result in unexpected behavior and power consumption in certain
situations, so it makes sense to avoid it if possible.

Refs #25