Add mini-panel configurations
For the main board, this is a 2-PCB version of the regular panel config.

For the transmission LED board, it is a 5-PCB version, but mostly a
proof of concept. It also adds an edge tab to work around a bug where
KiKit doesn't include the V-Cut there. Regardless, the resulting size
is too small for JLCPCB's V-Cut specs, and the boards are so simple
that there's really no benefit when doing hand-assembly.