Explicitly set the DTR signal when connecting
Valid DTR transitions are how the target device is able to detect when
the host is actually connected to it from a user application. Therefore,
consistent DTR state is a requirement to work correctly.

On Linux, the automatic DTR behavior appears to be correct.
On Windows, the automatic DTR behavior is somewhat inconsistent. If the
desktop app is the first/only app to connect to the target, it does not
appear to become set. However, connecting a terminal app will set/clear
it correctly. If a terminal app connects first, then the desktop app,
the desktop app will then implicitly control it correctly.

This change makes sure that we always explicitly set DTR when connecting
to the device, to avoid such issues.