Fix UID-handling code across all uses
The 96-bit UID from the MCU is read in a variety of places, for logging
purposes, system info purposes, and for USB device serial purposes. This
code was broken, in that it was not correctly reading the last word of
the UID.

This change fixes that, and uses the HAL wrapper functions to make it
more obvious how its working.

The ideal way to stringify the UID, in terms of format and endianness,
is unclear. So right now its more important that all the different
implementations produce the same output.

Fix #32