Implement support for USB device suspend/resume
When a USB host suspends a device, such as this one, it is supposed to
drop its power consumption to <2.5mA. This commit implements that
behavior, as well as the resume side.

To implement suspend, the device's main task is transitioned into a
special "suspend" state. This state, implemented via the common state
controller mechanism, shuts down all peripherals and puts the MCU into
"STOP" mode.  While in stop mode, the RTC is used to do a periodic
wakeup to kick the IWDG. Current testing shows that this means a
steady-state suspend current of ~0.7mA with periodic brief 1.3mA peaks.

In addition to the changes specific to this use case, a few other
related changes have been rolled in:

- Make the IWDG easier to disable by turning off the related HAL define

Sometimes for development purposes, we need to turn off the watchdog.
However, simply commenting out the code to enable it will cause the
WRPERR bit to get set in the Flash status register whenever watchdog
refresh code runs. This causes problem for any code that needs to work
with the flash, especially the bootloader. Therefore, all calls to the
IWDG API are wrapped or guarded and simply not run when the IWDG is
disabled.

- Keep track of the "default" home display state more globally, so any
  state that needs to return home (e.g. menu, remote, suspend) just
  needs to request a transition to STATE_HOME and not worry about
  specifics.

Fix #31