Changed settings API to use structs for related values
This is the first step in a series of planned changes to improve the way
settings are stored, for a variety of reasons including robustness,
future-proofing, and general cleanup.

Most of the calibration data stored in EEPROM settings can be grouped
intro bundles of related values that should be saved and loaded
together. Treating them all as independent key/value pairs increases the
chance of corruption and makes it slightly more awkward to pass them
around together. While a future commit may change the way these values
are persisted, first we must change the API used to manage them.

This change introduces a series of structs that bundle related values,
and changes the settings interface to work in terms of those structs. It
also makes a number of changes to the surrounding firmware code to deal
with this.

The most drastic of these surrounding changes involves changing the
calibration routines to separate measurements from saving results. This
slightly reduces code duplication, but moves more responsibility up to
the menu code that runs the calibration processes.