Redesigned the settings storage format
The new settings storage format is designed to improve the ability to
adapt to future changes, and to reduce the likelihood of partial writes
causing malformed data.

The first step was to split settings out into a series of distinct
"pages" of persisted data:
- Header page
- Sensor calibration page
- Target calibration page
- User settings page

For now, the header page is little more than a placeholder to detect a
fresh device, but it could be used to contain manufacturing data if the
need arises.

Likewise, the user settings page is currently empty. It will gain
contents when/if user settings or persisted state is added to the
device.

Each page has a designated region of memory, and a version number that
can be used if there are breaking format changes in the future.

Settings data will be stored one of two ways:
- Key/value pairs stored individually
- Value collections exposed as a struct

Value collections, such as anything externally represented with a
struct, are stored with a checksum to ensure that corruption from
partial writes is not cause for concern.

Additionally, a few changes to the EEPROM write code have been made to
improve performance and reduce watchdog issues:
- Data written in word-aligned blocks is now written as words, not bytes
- The IWDG is slowed down during the startup initialization because
  there are cases where a complete page initialization could be slow.