Add support for uncalibrated measurements
Being able to take measurements without first doing target calibration
is necessary to perform some calibration functions during the initial
setup. Since these are typically performed with aid from the desktop
app, this mode capability is only enabled when the desktop app is
connected.

The use cases for this involve slope calibration and out-of-band target
calibration.