Added USB-CDC command for wiping configuration data
This command is designed to be hard to send by accident, as the device
needs to be in remote control mode and the command arguments are also
device/firmware specific. It also resets the device after running.

Additionally, code was added to the settings EEPROM helper routines to
clear flash controller error flags prior to running. This removes the
need to explicitly workaround an unintentionally set WRP error flag in
the HAL code itself.