Added LED brightness detection to gain calibration
The transmission LED brightness level used for measuring maximum gain
was a hard-coded value. This was not ideal, since device-to-device
variations could make it not work so well.

To solve this problem, a measurement brightness detection routing has
been implemented. Now, prior to calibration max-vs-high gain, it walks a
sequence of LED brightness settings to find one that is closest to 75%
of the sensor's raw range.

Also reduced the LED cooldown delay, since it was likely needlessly long
before.

Ref #33