Initial implementation of diag remote control panel
This implements a GUI around the remote control commands currently
implemented on the device. It provides an easy way to control LED state
and the sensor read task.

To be truely useful for many testing/characterization use cases, a
number of modifications will be needed to the remote-controllable sensor
loop interface. However, right now this implements the interface as it
is.