Change remote control sensor diag mode to push results
This makes the desktop app's diagnostic control panel more closely
resemble the capabilities of the on-device diagnostics screen, and now
able to act as a suitable alternative.

The "get next reading" diagnostic command has now been removed, in favor
of simply pushing results as they arrive when the sensor task loop is
running.