Implement ADC oversampling to improve reading stability
This change enables 16x oversampling on the ADC peripheral, and converts
all the post-conversion calculations to operate in floating point (even
though the results are still integers). This change greatly improves the
stability of Vdda and Temperature readings, which will be neede for
device characterization.

If there is a need to later incorporate these into the sensor read loop,
then the internal representation of the temperature result will likely need
to be changed to floating point as well.

After these changes, the measurement routine takes approximately 5ms to
execute. If that ever needs to be a background operation with resepect
to the invoking task, it would be relatively easy to implement that.