Implemented basic device settings import/export capability
This is a minimal implementation of the feature, and uses a fairly plain
JSON file as its output format.

The export side is implemented with a QObject that "collects" all the
properties from the device and then serializes them to the file.

The import side is implemented with a dialog that loads the file,
displays a preview of its contents, and then allows for selective
import.

Ideally, there should be much better data encapsulation and separation
of concerns (especially on the import side). However doing so will first
require a major refactoring of how many of these properties are held
within the desktop app.  When/if a scriptable method for doing all of
this is implemented, that follow-on work should be done. For now,
however, this is probably good enough.