Reorganize directory structure of desktop app and add icon
This is the first step to incorporating deployment changes into the
project. It introduces a structure to the source tree that can be added
to, and also brings in an icon.

Then icon is currently based on a colored wireframe drawing of the
device, and is probably the best we can do without a graphic designer
being brought in.