Added handling of USB suspend/resume race conditions
If suspend and resume events happen in quick succession, the device may
receive a resume event while it is in the middle of processing a
suspend. When this happens, the device can get stuck in suspend until it
is power-cycled.

The fix is to check for this case in two places:
First, a check is added right before the code that actually starts the
process of going into low-power mode.
Second, a check is added to the RTC wakeup routine that runs
periodically during suspend to kick the watchdog.

This whole case is relatively rare, and has only been observed after
doing a UF2 firmware update from a Linux host, but it is consistently
reproducible. Therefore, it should be handled, just in case it can
happen under other circumstances.

Fix #38