Implement USB HID option for device output
In addition to the existing USB-CDC interface for sending measurements
to a host PC, this adds the ability to also use a virtual USB-HID device
to send them.  The idea is that, if the primary use case is to simply
"paste measurements into a spreadsheet", it would be a lot easier if the
device could simply "type them in" without needing intermediate software
to handle the process.

This feature is off by default, but can be enabled in device settings.
It has the following configuration options:
- Enabled (yes/no)
- Output format (e.g. "0.37" vs "R+0.37D")
- Output seperator (e.g. enter, tab, comma, space)

When disabled, the device will continue to just present itself as a CDC
device class. When enabled, the device will present itself as both CDC
and HID. Toggling the setting will cause the device to
disconnect/reconnect from the host to implement this change.

Furthermore, when a host application is connected to the CDC device
(i.e. the DTR flag is active), writes to the HID interface are
automatically inhibited.

For the implementation of this feature, the USBD task code was moved to
its own source file, and HID processing was added into the USBD task.
This is because HID writes are started from the main task, but continued
via successive calls to USBD-task callback functions.

Fix #39