Adjust I2C bus timings to use larger and more explicit rise/fall times
While the previous values worked fine on all of the prototype devices to
date, they did not work correctly with the TSL2591 sensors from the reel
slated for use on production devices. As such, the final batch of
pre-production devices (that used parts from this reel) failed to start
correctly.

Making this change appears to correct the issue.

It should also be noted that all of the STM-provided example projects
that use I2C also have values similar to what we have with this change.