Initial implementation of measurement capture table
Added a table to the measurement tab to capture readings.
The latest reading can be added automatically and/or explicitly.
The whole table can be copied or cleared.
Table auto-updates only happen when the measurement tab is active.

There are still a lot of finishing touches for hotkeys and various
interactions for table management (e.g. inserting rows, deleting rows,
context menus, copy selected regions, etc) but this is a good first
step.