Add fancier edit actions to the application
This adds the edit menu, with its cut/copy/paste/delete actions, as well
as giving more explicit consideration to how these actions are to be
handled by the application.
There may still be room for improvement or alternate decisions
in regards to the measurement table, but this is at least what we need
to have these at all.