Reoganize calibration menu idle light control
Make all control over the idle state of the measurement lights exclusive
to the menu code surrounding the calls to the calibration functions, and
only turn on those lights when the device is waiting for the user to
take a measurement.

This slightly reduces the amount of time the lights are on in that state
prior to a measurement.

Also, for these interactions, the brightness is fixed for consistency.
The configured value is not used here.