Various minor PCB fixups
Changed C1 footprint from an 0402 resistor to an 0402 capacitor. This
change shoulds be metadata only, or close to it, but makes the BOM data
less confusing.

Updated PCB metadata from schematic, which should pull in all of the BOM
manufacturer/vendor changes that were made when procuring components for
the production run.  It should also pull in any net name changes due to
KiCad updates.